complete order, add policies and enum using

+++ app/Http/Controllers/ProductController.php
@@ -2,27 +2,31 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\CreateProductRequest;
+use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Http\Request;
 
 class ProductController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ProductResource::collection(Product::all());
+        Gate::authorize('index', Product::class);
+        return ['products' => ProductResource::collection(Product::all())];
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(CreateProductRequest $request)
+    public function store(StoreProductRequest $request)
     {
+        Gate::authorize('store', Product::class);
         $product = $request->validated();
         $product['image'] = $request->file('image')->storeAs('images', $product['title'], 'public');
         return new ProductResource(Product::create($product));
@@ -33,6 +37,7 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        Gate::authorize('show', Product::class);
         return new ProductResource($product);
     }
 
@@ -41,6 +46,7 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, Product $product)
     {
+        Gate::authorize('update', Product::class);
         $product->update($request->validated());
         return new ProductResource($product);
     }
@@ -50,6 +56,7 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('destroy', Product::class);
         Storage::disk('public')->delete($product['image']);
         $product->delete();
         return ['Product deleted'];
